split APIClient
